Add Pleiades part 1 -> Pleiades part 2 migration
In Pleiades part 2 we will be adding a new parameter:
MonitoredTokenAddresses for cross-bridge balance checks. The chain will
panic on the first EndBlocker execution if we do not add a migration for the
Params to add a default value and a ParamSetPair for this new param.

We also added a genesis value, but it remains to be seen if that
requires migration logic.